+++ pkg/cortex/modules.go
@@ -21,6 +21,7 @@ import (
 	"github.com/cortexproject/cortex/pkg/alertmanager"
 	"github.com/cortexproject/cortex/pkg/chunk"
 	"github.com/cortexproject/cortex/pkg/chunk/storage"
+	"github.com/cortexproject/cortex/pkg/compactor"
 	"github.com/cortexproject/cortex/pkg/configs/api"
 	"github.com/cortexproject/cortex/pkg/configs/db"
 	"github.com/cortexproject/cortex/pkg/distributor"
@@ -53,6 +54,7 @@ const (
 	Ruler
 	Configs
 	AlertManager
+	Compactor
 	All
 )
 
@@ -84,6 +86,8 @@ func (m moduleName) String() string {
 		return "configs"
 	case AlertManager:
 		return "alertmanager"
+	case Compactor:
+		return "compactor"
 	case All:
 		return "all"
 	default:
@@ -128,6 +132,9 @@ func (m *moduleName) Set(s string) error {
 		return nil
 	case "alertmanager":
 		*m = AlertManager
+		return nil
+	case "compactor":
+		*m = Compactor
 		return nil
 	case "all":
 		*m = All
@@ -458,6 +465,16 @@ func (t *Cortex) stopAlertmanager() error {
 	return nil
 }
 
+func (t *Cortex) initCompactor(cfg *Config) (err error) {
+	t.compactor, err = compactor.NewCompactor(cfg.Compactor, cfg.TSDB, util.Logger, prometheus.DefaultRegisterer)
+	return err
+}
+
+func (t *Cortex) stopCompactor() error {
+	t.compactor.Shutdown()
+	return nil
+}
+
 type module struct {
 	deps []moduleName
 	init func(t *Cortex, cfg *Config) error
@@ -537,6 +554,12 @@ var modules = map[moduleName]module{
 		deps: []moduleName{Server},
 		init: (*Cortex).initAlertmanager,
 		stop: (*Cortex).stopAlertmanager,
+	},
+
+	Compactor: {
+		deps: []moduleName{Server},
+		init: (*Cortex).initCompactor,
+		stop: (*Cortex).stopCompactor,
 	},
 
 	All: {